mtd: nand: Driver for Evatronix NAND flash controller
Hi,

Here's the patch again, same email that I sent to the mtd list.

/Ricard

[PATCH RFC]: mtd: nand: Driver for Evatronix NAND flash controller

From: Ricard Wanderlof <ricardw@axis.com>

This is a request for comments for a driver for the Evatronix NANDFLASH-CTRL
IP (version 1.15). It is designed to be applied to Linux 3.16 or the current
linux-mtd tree.

Currently, it supports one instance of the NANDFLASH-CTRL IP, with up
to two connected NAND flash chips, one per bank. (The IP itself is very
configurable, but the intent has been to get something working for a
specific configuration.)

Signed-off-by: Ricard Wanderlof <ricardw@axis.com>